Change the format of the bytecode function
Now it does much less explicit parameter handling; instead, it's
"built in", and makes a distinction between required, optional,
and rest parameters. Later on there'll be a need for a "param"
block at the start of the function, and special instructions to
check whether the corresponding argument was passed. But not now.