Implement non-tail recursion elimination
Wow, that was way too difficult, considering the transformation
is actually fairly straightforward. It feels like there are not
enough solid abstractions for querying/transforming target code
(which is maybe a good thing to address separately anyway).

Secondly, I think the real way to do this is during the normal
syntax-driven compilation, not as a separate post-processing
pass. But let's explore abstractions for target code first, and
then change it to happen earlier.